refactor: コンポーネント化->AIたくさん使用
