+++ PrefeFortune/Views/TouristCard/TouristCardView.swift
